+++ datanode/api/trading_data_v2.go
@@ -543,8 +543,16 @@ func (t *tradingDataServiceV2) ListCandleData(ctx context.Context, req *v2.ListC
 		return nil, errors.New("sql candle service not available")
 	}
 
-	from := vegatime.UnixNano(req.FromTimestamp)
-	to := vegatime.UnixNano(req.ToTimestamp)
+	var from, to *time.Time
+	if req.FromTimestamp != 0 {
+		fromTimestamp := vegatime.UnixNano(req.FromTimestamp)
+		from = &fromTimestamp
+	}
+
+	if req.ToTimestamp != 0 {
+		toTimestamp := vegatime.UnixNano(req.ToTimestamp)
+		to = &toTimestamp
+	}
 
 	pagination := entities.CursorPagination{}
 	if req.Pagination != nil {
@@ -554,7 +562,7 @@ func (t *tradingDataServiceV2) ListCandleData(ctx context.Context, req *v2.ListC
 		}
 	}
 
-	candles, pageInfo, err := t.candleService.GetCandleDataForTimeSpan(ctx, req.CandleId, &from, &to, pagination)
+	candles, pageInfo, err := t.candleService.GetCandleDataForTimeSpan(ctx, req.CandleId, from, to, pagination)
 	if err != nil {
 		return nil, apiError(codes.Internal, ErrCandleServiceGetCandleData, err)
 	}